Adicionado um Helper para manipular as respostas do serviços de acordo com o StatusCode retornado.

+++ ExpenseApi/ExpenseApi.Service/Service/UserService.cs
@@ -24,21 +24,31 @@ namespace ExpenseApi.Service.Service
 
         public async Task<ServiceResult<User>> CreateAsync(User user)
         {
+            var userExists = (await _repository.FindAsync(x => x.Email == user.Email))?.FirstOrDefault();
+
+            if (userExists != null)
+                return ServiceResult<User>.CreateInvalidResult($"Já existe um usuário cadastrado com este email: {user.Email}");
+
             user.Id = ObjectId.GenerateNewId();
             user.Password = _passwordHasher.HashPassword(user.Password);
 
             var entity = await _repository.CreateAsync(user);
             entity.ClearPassword();
-            return ServiceResult<User>.CreateValidResult();
+            return ServiceResult<User>.CreateValidResult(entity);
         }
         public async Task<ServiceResult<User>> UpdateAsync(User user, bool isUpdatePassword = true)
         {
-            if(isUpdatePassword)
+            if (isUpdatePassword)
                 user.Password = _passwordHasher.HashPassword(user.Password);
 
-            var entity = await _repository.UpdateAsync(user);
-            entity.ClearPassword();
-            return ServiceResult<User>.CreateValidResult(entity);
+            var entity = await _repository.GetByIdAsync(user.Id);
+
+            if (entity == null)
+                return ServiceResult<User>.CreateInvalidResult("Registro não encontrado.");
+
+            var result = await _repository.UpdateAsync(user);
+            result.ClearPassword();
+            return ServiceResult<User>.CreateValidResult(result);
         }
 
         public async Task<ServiceResult<User>> UpdatePasswordAsync(string userId, string oldPassword, string newPassword)
@@ -59,40 +69,49 @@ namespace ExpenseApi.Service.Service
             return ServiceResult<User>.CreateValidResult(result);
         }
 
-        public async Task DeleteAsync(string id)
+        public async Task<ServiceResult<bool>> DeleteAsync(string id)
         {
+            var entity = await _repository.GetByIdAsync(ObjectId.Parse(id));
+
+            if (entity == null)
+                return ServiceResult<bool>.CreateInvalidResult("Registro não encontrado.");
+
             await _repository.DeleteAsync(new ObjectId(id));
+            return ServiceResult<bool>.CreateValidResult(true);
         }
 
-        public async Task<ServiceResult<List<User>>> FindAsync(Expression<Func<User, bool>> filterExpression)
+        public async Task<ServiceResult<List<User>>> FindAsync(Expression<Func<User, bool>> filterExpression, bool clearPassword = true)
         {
             var result = await _repository.FindAsync(filterExpression);
 
-            result.ForEach(user => { user.ClearPassword(); });
+            if(clearPassword)
+                result.ForEach(user => { user.ClearPassword(); });
 
-            return ServiceResult<List<User>>.CreateValidResult();
+            return ServiceResult<List<User>>.CreateValidResult(result);
         }
 
-        public async Task<ServiceResult<List<User>>> GetAllAsync()
+        public async Task<ServiceResult<List<User>>> GetAllAsync(bool clearPassword = true)
         {
             var result = await _repository.FindAsync(_ => true);
+            
+            if (clearPassword)
+                result.ForEach(user => { user.ClearPassword(); });
 
-            result.ForEach(user => { user.ClearPassword(); });
-
-            return ServiceResult<List<User>>.CreateValidResult();
+            return ServiceResult<List<User>>.CreateValidResult(result);
         }
 
-        public async Task<ServiceResult<User>> GetByIdAsync(string id)
+        public async Task<ServiceResult<User>> GetByIdAsync(string id, bool clearPassword = true)
         {
             var entity = await _repository.GetByIdAsync(new ObjectId(id));
             if (entity == null)
                 return ServiceResult<User>.CreateInvalidResult("Registro não encontrado.");
 
-            entity.ClearPassword();
+            if (clearPassword)
+                entity.ClearPassword();
 
             return ServiceResult<User>.CreateValidResult(entity);
         }
 
-        
+
     }
 }
